backend/requirements-core.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
requests
llama-index-core
llama-index-readers-file
llama-index-readers-web
html2text
pypdf
docx2txt
llama-index-llms-anthropic
